Unit test fixed
