+++ package.json
@@ -14,7 +14,6 @@
     "react-router-dom": "^6.14.2",
     "react-scripts": "5.0.1",
     "redux-logger": "^3.0.6",
-    "redux-mock-store": "^1.5.4",
     "redux-thunk": "^2.4.2",
     "uuid": "^9.0.0",
     "web-vitals": "^2.1.4"
@@ -51,7 +50,7 @@
     "@babel/preset-act": "file:babel/preset-react@7.xre@7.x",
     "@babel/preset-react": "^7.14.5",
     "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "latest",
+    "@testing-library/react": "^14.0.0",
     "babel-jest": "^27.5.1",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
@@ -60,6 +59,7 @@
     "eslint-plugin-react": "^7.33.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "jest": "^27.5.1",
-    "react-test-renderer": "^18.2.0"
+    "react-test-renderer": "^18.2.0",
+    "redux-mock-store": "^1.5.4"
   }
 }
